--- process.md
@@ -1,13 +1,0 @@
-# the-super-tiny-compiler
-
-> forked from [jamiebuilds/the-super-tiny-compiler](https://github.com/jamiebuilds/the-super-tiny-compiler)
-
-
-- 测试
-
-Most compilers break down into three primary stages: Parsing, Transformation, and Code Generation:   
-- *Parsing* is taking raw code and turning it into a more abstract representation of the code.
-- *Transformation* takes this abstract representation and manipulates to do whatever the compiler wants it to.
-- *Code Generation* takes the transformed representation of the code andturns it into new code.
-
-
